[merge] back/franchise/master to back/master

See merge request s11-fintech-finance-sub1/S11P21A601!151

+++ backend/franchise/src/main/java/kr/co/olivepay/franchise/dto/res/FranchiseBasicRes.java
@@ -18,8 +18,8 @@ import lombok.Builder;
 public record FranchiseBasicRes (
 	Long franchiseId,
 	String franchiseName,
-	Float latitude,
-	Float longitude,
+	Double latitude,
+	Double longitude,
 	String category,
 	Integer likes,
 	Integer coupons,
